split up the sites class into logic and view classes
